# keylock: make KeyLock generic and improve documentation